+++ infra-cdk/bin/fast-cdk.ts
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 import * as cdk from "aws-cdk-lib"
-import { GaspMainStack } from "../lib/gasp-main-stack"
+import { FastMainStack } from "../lib/fast-main-stack"
 import { ConfigManager } from "../lib/utils/config-manager"
 
 // Load configuration using ConfigManager
@@ -12,7 +12,7 @@ const props = configManager.getProps()
 const app = new cdk.App()
 
 // Deploy the new Amplify-based stack that solves the circular dependency
-const amplifyStack = new GaspMainStack(app, props.stack_name_base, {
+const amplifyStack = new FastMainStack(app, props.stack_name_base, {
   config: props,
   env: { 
     account: process.env.CDK_DEFAULT_ACCOUNT,