+++ infra-cdk/test/fast-cdk.test.ts
@@ -1,13 +1,13 @@
 // import * as cdk from 'aws-cdk-lib';
 // import { Template } from 'aws-cdk-lib/assertions';
-// import * as GaspCdk from '../lib/gasp-cdk-stack';
+// import * as FastCdk from '../lib/fast-cdk-stack';
 
 // example test. To run these tests, uncomment this file along with the
-// example resource in lib/gasp-cdk-stack.ts
+// example resource in lib/fast-cdk-stack.ts
 test('SQS Queue Created', () => {
 //   const app = new cdk.App();
 //     // WHEN
-//   const stack = new GaspCdk.GaspCdkStack(app, 'MyTestStack');
+//   const stack = new FastCdk.FastCdkStack(app, 'MyTestStack');
 //     // THEN
 //   const template = Template.fromStack(stack);